stats: Move Counter* and Gauge* types into new file counter.go.
Move Duration* types into duration.go.

This is a mere move of code and no change in functionality.

By splitting out the code, I hope it will be easier to maintain it.
