+++ go/stats/export.go
@@ -36,7 +36,6 @@ import (
 	"sync"
 	"time"
 
-	"vitess.io/vitess/go/sync2"
 	"vitess.io/vitess/go/vt/log"
 )
 
@@ -194,52 +193,6 @@ type FloatFunc func() float64
 // String is the implementation of expvar.var
 func (f FloatFunc) String() string {
 	return strconv.FormatFloat(f(), 'g', -1, 64)
-}
-
-// Duration exports a time.Duration
-type Duration struct {
-	i sync2.AtomicDuration
-}
-
-// NewDuration returns a new Duration
-func NewDuration(name string) *Duration {
-	v := new(Duration)
-	publish(name, v)
-	return v
-}
-
-// Add adds the provided value to the Duration
-func (v *Duration) Add(delta time.Duration) {
-	v.i.Add(delta)
-}
-
-// Set sets the value
-func (v *Duration) Set(value time.Duration) {
-	v.i.Set(value)
-}
-
-// Get returns the value
-func (v *Duration) Get() time.Duration {
-	return v.i.Get()
-}
-
-// String is the implementation of expvar.var
-func (v *Duration) String() string {
-	return strconv.FormatInt(int64(v.i.Get()), 10)
-}
-
-// DurationFunc converts a function that returns
-// an time.Duration as an expvar.
-type DurationFunc func() time.Duration
-
-// String is the implementation of expvar.var
-func (f DurationFunc) String() string {
-	return strconv.FormatInt(int64(f()), 10)
-}
-
-// FloatVal is the implementation of MetricFunc
-func (f DurationFunc) FloatVal() float64 {
-	return f().Seconds()
 }
 
 // String is expvar.String+Get+hook
